Last years code changes

+++ secret-santa-sharp/secret-santa-sharp/Program.cs
@@ -11,36 +11,41 @@ namespace secret_santa_sharp
    {
       static void Main(string[] args)
       {
-         Console.WriteLine("Hello World!");
-
+         var dryRun = !(args.Length == 1 && args[0] == "hohoho");
+         Console.WriteLine($"Running secret santa matcher! {(dryRun ? "Dry Run" : "")}");
          var pickSettings = PickingSettings.LoadSettings();
+
          MatchSantas(pickSettings);
 
-         File.WriteAllText(
-            $"ps-{DateTime.Now.ToString("yyyy-dd-M--HH-mm-ss")}.json",
-            JsonConvert.SerializeObject(pickSettings, Formatting.Indented));
+         if (!dryRun)
+         {
+            File.WriteAllText(
+               $"ps-{DateTime.Now.ToString("yyyy-dd-M--HH-mm-ss")}.json",
+               JsonConvert.SerializeObject(pickSettings, Formatting.Indented));
+         }
 
          TwilioClient.Init(pickSettings.TwilioSid, pickSettings.TwilioAuthToken);
-         SendMessages(pickSettings);
+         SendMessages(pickSettings, dryRun);
       }
 
-      private static void SendMessages(PickingSettings ps)
+      private static void SendMessages(PickingSettings ps, bool dryRun)
       {
          foreach (var matchId in ps.Pairing)
          {
             var santa = ps.GetSantaById(matchId.Key);
             var match = ps.GetSantaById(matchId.Value);
-            
-            var msg = $"Hello {santa.Name}, sorry about that the elves got confused. You have matched with {match.Name}! Merry Christmas!";
+
+            var msg = $"Hello {santa.Name}, you will be secret santa for {match.Name}! Merry Christmas!";
             Console.WriteLine(msg);
             Console.WriteLine($"Number={santa.PhoneNumber}");
 
-            SendText(msg, ps.SenderNumber, santa.PhoneNumber);
+            if (!dryRun) SendText(msg, ps.SenderNumber, santa.PhoneNumber);
          }
       }
 
       private static void SendText(string body, string fromNumber, string toNumber)
       {
+         Console.WriteLine($"Send text to {toNumber}");
          var msg = MessageResource.Create(
                body: body,
                from: new Twilio.Types.PhoneNumber(fromNumber),
@@ -51,7 +56,7 @@ namespace secret_santa_sharp
 
       private static void MatchSantas(PickingSettings pickSettings)
       {
-         var ran = new Random((int)DateTime.Now.ToBinary());
+         var ran = new Random((int)DateTime.Now.Ticks);
          var availableSantas = pickSettings.Santas.ToDictionary(k => k.Id);
          foreach (var santa in pickSettings.Santas)
          {
